move method from node utils to all utils
