+++ src/transcription/node/utils.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosInstance } from 'axios'
+import { AxiosInstance } from 'axios'
 import FormData from 'form-data'
 import { createReadStream, existsSync } from 'node:fs'
 import path from 'node:path'
@@ -155,37 +155,4 @@ export const getTranscriptionLocalFileSource = async (
   const signFileRequestUrl = `/organizations/${organizationName}/file_tokens/${fileName}`
   const { url } = await signFile(axiosInstance, signFileRequestUrl)
   return url
-}
-
-/**
- * Checks if a URL is likely to be downloadable.
- *
- * @param url - The URL to check.
- * @returns A Promise that resolves to true if the URL appears to be downloadable; otherwise, false.
- */
-export const checkIfUrlDownloadable = async (url: string): Promise<boolean> => {
-  try {
-    const response = await axios.get(url, {
-      headers: {
-        Range: 'bytes=0-0',
-      },
-      responseType: 'stream',
-      maxRedirects: 5,
-    })
-
-    if (response.status === 206 || response.status === 200) {
-      const contentDisposition = response.headers['content-disposition'] || ''
-      if (contentDisposition.toLowerCase().includes('attachment')) {
-        return true
-      }
-
-      const contentType = response.headers['content-type'] || ''
-      if (!contentType.toLowerCase().startsWith('text/html')) {
-        return true
-      }
-    }
-  } catch (error) {
-    return false
-  }
-  return false
 }
